requirements.txt
altair==5.5.0
annotated-types==0.7.0
appdirs==1.4.4
appnope==0.1.4
art==6.2
asciichartpy==1.5.25
asttokens==2.4.1
attrs==24.3.0
av==13.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
comm==0.2.2
confuse==2.0.1
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
executing==2.0.1
fonttools==4.54.1
freetype-py==2.3.0
gitdb==4.0.11
GitPython==3.1.43
html2rml==0.3.2
icalendar==6.1.0
idna==3.7
imageio==2.36.0
ipydex==0.18.0
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.1
mdurl==0.1.2
mpfshell==0.9.3
nanoid==2.0.0
narwhals==1.19.1
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.0.0
numpy-stl==3.1.2
ome-types==0.5.3
opencv-python==4.10.0.84
owlready2==0.47
packaging==24.1
paho-mqtt==2.1.0
pandas==2.2.2
parso==0.8.4
pdoc==14.5.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
plotext==5.2.8
plumbum==1.8.3
prompt_toolkit==3.0.47
protobuf==5.29.2
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1725737856345/work
ptpython==3.0.29
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pycairo==1.27.0
pydantic==2.10.4
pydantic-compat==0.1.2
pydantic_core==2.27.2
pydeck==0.9.1
pyglet==2.0.18
Pygments==2.18.0
PyOpenGL==3.1.0
pypandoc==1.13
pyparsing==3.1.4
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyqtgraph==0.13.7
pyrender==0.1.45
pyserial==3.5
pyte==0.8.2
python-dateutil==2.9.0.post0
python-utils==3.9.0
pytz==2024.1
pyudev==0.24.3
PyYAML==6.0.1
pyzmq==26.2.0
qtconsole==5.6.0
QtPy==2.4.1
rdflib==7.1.1
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
rich==13.7.1
rlPyCairo==0.3.0
rpds-py==0.22.3
rpyc==6.0.1
rshell==0.0.33
schedule==1.2.2
scipy==1.14.1
setuptools==70.1.1
six==1.16.0
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.41.1
tabulate==0.9.0
tenacity==9.0.0
textual==0.71.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
trimesh==4.5.1
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
xsdata==24.3.1
yamlpyowl==0.2.1
